MSVC x64 support; fixed implicit cast warns
The most annoying thing about MSVC x64 is that "unsigned long" is still
32-bits, which causes a lot of problems with MPIR and mpz_get_ui. We add
types gmp_ui and gmp_si to deal with this.

+++ arith/fieldquadratic.c
@@ -126,7 +126,7 @@ static int fq_snprint(char *s, size_t n, element_ptr e) {
   clip_sub();
   status = snprintf(s + result, left, "]");
   if (status < 0) return status;
-  return result + status;
+  return (int)(result + status);
   #undef clip_sub
 }
 
@@ -153,7 +153,7 @@ static int fq_set_str(element_ptr e, const char *s, int base) {
   if (*cp++ != ',') return 0;
   cp += element_set_str(p->y, cp, base);
   if (*cp++ != ']') return 0;
-  return cp - s;
+  return (int)(cp - s);
 }
 
 static int fq_sign(element_ptr n) {
